requirements.txt
alembic==0.9.2
blinker==1.4
click==6.7
coverage==4.3.4
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Email==1.4.4
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.4
Flask-Moment==0.5.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
Flask-SocketIO==2.8.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
mysqlclient==1.3.10
numpy==1.13.0
python-dateutil==2.6.0
python-editor==1.0.3
python-engineio==1.5.4
python-socketio==1.7.5
six==1.10.0
SQLAlchemy==1.1.10
stomp.py==4.1.18
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
